Fixed compiler warnings related to implicit float/double casts

+++ libraries/audio/src/AudioRingBuffer.cpp
@@ -211,7 +211,7 @@ float AudioRingBuffer::getFrameLoudness(const int16_t* frameStart) const {
     const int16_t* _bufferLastAt = _buffer + _bufferLength - 1;
 
     for (int i = 0; i < _numFrameSamples; ++i) {
-        loudness += std::abs(*sampleAt);
+        loudness += (float) std::abs(*sampleAt);
         sampleAt = sampleAt == _bufferLastAt ? _buffer : sampleAt + 1;
     }
     loudness /= _numFrameSamples;
